Make database use int64 instead of int32 for meeting id

+++ jsonmodel.go
@@ -4,14 +4,14 @@ type JSONEntry struct {
 	ID           uint   `json:"id"`
 	Description  string `json:"description"`
 	CreationDate int64  `json:"creation_date"`
-	MeetingID    int32  `json:"meeting_id"`
+	MeetingID    int64  `json:"meeting_id"`
 	Passcode     string `json:"passcode"`
 	PasscodeHash string `json:"passcode_hash"`
 }
 
 type JSONCreateEntry struct {
 	Description  string `json:"description"`
-	MeetingID    int32  `json:"meeting_id"`
+	MeetingID    int64  `json:"meeting_id"`
 	Passcode     string `json:"passcode"`
 	PasscodeHash string `json:"passcode_hash"`
 }
